v2/guest/syscall: Account for optval alignment in setsockopt
We have no means to determine the actual alignment of optval if we
receive the call via opaque `syscall` method on the guest side and
there is no way to determine it on the host side.
Align it to `align_of::<usize>()` as a precaution, which is probably the
best we can do here.

Signed-off-by: Roman Volosatovs <roman@profian.com>