refactor: vendor libc for compiling on x86_64-unknown-none
The libc crate naturally doesn't expose any Linux-specific
items when compiling for the x86_64-unknown-none target,
but we need them anyway, so in order to allow compiling
for that target we provide our own definitions.

If we ever expand to supporting other platforms we can
reintroduce libc with cfg workarounds.
Furthermore, someday the libc crate might be split into
multiple platform-specific crates, at which point we may
be able to get rid of the definitions here and rely on
those crates directly.

Closes https://github.com/enarx/sallyport/issues/110